backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
yfinance==0.2.40
stripe==8.0.0
